refactor(identity): factor shared caller + axum auth into b00t-c0re-identity (SP4-07)
CallerIdentity + JwksVerifier -> b00t-c0re-identity::caller (SP3-01 impl).
identity_middleware + IdentityLayerState -> b00t-c0re-identity::axum_auth,
gated by the crate's new `axum` feature. b00t-mcp's identity.rs / http_auth.rs
become re-export shims that keep SP3-02a's tests running verbatim (CP-4 auth
parity). Every backend MCP server can now depend on b00t-c0re-identity with
features=["axum"] and run byte-identical bearer verification.
