+++ b00t-mcp/Cargo.toml
@@ -37,7 +37,7 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 # b00t-mcp specific dependencies
 b00t-cli = { workspace = true }
 b00t-c0re-lib = { workspace = true }
-b00t-c0re-identity = { workspace = true }
+b00t-c0re-identity = { workspace = true, features = ["axum"] }
 b00t-c0re-ledgrrr = { workspace = true }
 capability-forge = { path = "../capability-forge" }
 ufo-types = { workspace = true }
